Merged revision(s) 2077, 2079-2080, 2082, 2089, 2091, 2096 from vwf/development: Updated Humvee from Third Track with texture paths made relative. 2012-07-02 version.
........
Articulation demo for 2012-06-28 version. Requires the node3.vwf "frame" patch.

........
Articulation demo for 2012-07-02 version. Requires the node3.vwf "frame" patch.

........
Merged branches/frames for the rotateBy() "scaled" frame of reference to support the Humvee articulation behavior.

........
Updating worldTransform to use this.parent.worldTransform
........
Adding origin information to pickInfo
........
Work around the UTF-8 BOM for YAML too, and restrict the search to the start of the string, which is presumably faster.

........


git-svn-id: http://svn.virtualworldframework.com/vwf/integration@2908 775f9103-8eee-4f94-88be-65e0ca6d8b75